--- src/constants/cupcakes.ts
@@ -1,102 +1,0 @@
-// List of possible cupcake flavors. These flavors can be used to customize your cupcake orders.
-export const CUPCAKE_FLAVORS = [
-  "Vanilla", // A classic, light, and sweet flavor with a hint of vanilla.
-  "Chocolate", // A rich and decadent chocolate flavor.
-  "Red Velvet", // A soft, moist cake with a hint of cocoa and a vibrant red color.
-  "Carrot", // A spiced cake made with grated carrots and often includes nuts or raisins.
-  "Lemon", // A zesty and refreshing lemon-flavored cupcake.
-  "Strawberry", // A fruity and sweet flavor with fresh strawberry essence.
-  "Marble", // A swirl of vanilla and chocolate for a marble effect and dual flavors.
-  "Funfetti", // A colorful cupcake filled with sprinkles for a fun, festive look.
-  "Coffee", // A rich and robust flavor inspired by coffee beans.
-  "Coconut", // A tropical flavor with a subtle coconut essence.
-  "Cookies and Cream", // A sweet flavor with crushed cookies mixed into the batter.
-  "Peanut Butter", // A nutty and creamy cupcake made with peanut butter.
-  "Salted Caramel", // A sweet and salty flavor made with caramel and a touch of salt.
-  "Mocha", // A coffee and chocolate blend for a rich, mocha-inspired flavor.
-  "Matcha", // A delicate, earthy flavor made with powdered green tea.
-  "Banana", // A moist cupcake with the sweet and fruity taste of bananas.
-  "Almond", // A nutty flavor with a rich, aromatic almond essence.
-  "Chocolate Mint", // A refreshing combination of mint and chocolate.
-  "Raspberry", // A tart and sweet flavor made with fresh raspberries.
-  "Tiramisu", // A coffee-flavored cupcake inspired by the classic Italian dessert.
-  "Mint Chocolate Chip", // A refreshing mint flavor with chocolate chips mixed in.
-  "Maple Bacon", // A savory-sweet flavor with maple syrup and crispy bacon bits.
-  "Churro", // A cinnamon-sugar flavored cupcake inspired by the classic churro.
-]
-
-// List of frosting options for cupcakes. These frostings can be used to decorate and flavor cupcakes.
-export const CUPCAKE_FROSTINGS = [
-  "Buttercream", // A rich, smooth frosting made with butter, sugar, and vanilla.
-  "Cream Cheese Frosting", // A tangy, smooth frosting made with cream cheese and sugar.
-  "Chocolate Frosting", // A rich, chocolatey frosting for any chocolate cupcake.
-  "Whipped Cream", // A light, fluffy frosting made with whipped cream.
-  "Ganache", // A glossy, rich frosting made from chocolate and cream.
-  "Swiss Meringue Buttercream", // A silky, smooth buttercream made with egg whites, sugar, and butter.
-  "Italian Meringue Buttercream", // A smooth, stable buttercream made with egg whites and sugar syrup.
-  "Peanut Butter Frosting", // A creamy frosting made with peanut butter for a nutty sweetness.
-  "Caramel Frosting", // A rich, sweet frosting made with caramelized sugar.
-  "Salted Caramel Frosting", // A frosting with a sweet and salty flavor combination.
-  "Coconut Frosting", // A smooth frosting with the rich flavor of coconut.
-  "Strawberry Frosting", // A sweet frosting made with fresh strawberries.
-  "Lemon Frosting", // A tangy, citrusy frosting made with lemon zest.
-  "Mint Frosting", // A refreshing frosting with mint flavor.
-  "Coffee Frosting", // A rich, coffee-flavored frosting.
-  "Marshmallow Frosting", // A light, fluffy frosting made with marshmallows.
-  "Mocha Frosting", // A blend of coffee and chocolate in a smooth frosting.
-  "Raspberry Frosting", // A fruity, sweet frosting made with raspberries.
-  "Churro Frosting", // A spiced frosting with cinnamon and sugar, inspired by churros.
-  "Chocolate Hazelnut Frosting", // A creamy, nutty frosting with chocolate and hazelnut flavors.
-]
-
-// List of possible cupcake toppings. These toppings can be added for extra flavor and decoration.
-export const CUPCAKE_TOPPINGS = [
-  "Sprinkles", // Colorful sprinkles to add a fun, festive touch.
-  "Chopped Nuts", // Chopped nuts like almonds, walnuts, or pecans for a crunchy texture.
-  "Fresh Fruit (Strawberries, Blueberries, etc.)", // Fresh fruit like strawberries, blueberries, or raspberries for a fruity topping.
-  "Chocolate Chips", // Classic chocolate chips for extra sweetness.
-  "Mini Marshmallows", // Tiny marshmallows for a soft, chewy topping.
-  "Caramel Drizzle", // A sweet drizzle of caramel sauce for added richness.
-  "Shredded Coconut", // A tropical topping made with shredded coconut.
-  "Oreos", // Crushed Oreos for a chocolatey crunch.
-  "Edible Flowers", // Beautiful, edible flowers to add a touch of elegance.
-  "Meringue Kisses", // Light, crispy meringue kisses for a delicate topping.
-  "Chocolate Shavings", // Thin shavings of chocolate for a decorative finish.
-  "Graham Cracker Crumbs", // Crushed graham crackers for a crunchy topping.
-  "Candy Pieces", // Colorful candy pieces like M&Ms or Skittles for a playful topping.
-  "Peanut Butter Cups", // Chopped peanut butter cups for a rich, nutty topping.
-  "Cotton Candy", // Sweet and fluffy cotton candy for a whimsical touch.
-  "Rainbow Dust", // Shimmering dust for a colorful, sparkling effect.
-  "Whipped Cream", // Light and fluffy whipped cream for a creamy topping.
-  "Candied Ginger", // Sweet, candied ginger for a spicy-sweet topping.
-  "Crushed Cookies", // Crushed cookies like Oreos or chocolate chip cookies for texture.
-  "Butter Toffee", // Sweet, crunchy butter toffee for an indulgent topping.
-  "Gold Leaf", // Elegant gold leaf for a luxurious touch.
-  "Pretzels", // Crispy, salty pretzels for a sweet-salty contrast.
-  "Cinnamon Sugar", // A blend of cinnamon and sugar for a sweet, spiced topping.
-]
-
-// List of possible cupcake fillings. These fillings can be added to cupcakes for extra flavor and texture.
-export const CUPCAKE_FILLINGS = [
-  "Vanilla Custard", // A creamy, rich custard filling with a vanilla flavor.
-  "Chocolate Ganache", // A thick, rich chocolate ganache filling.
-  "Strawberry Jam", // A sweet and fruity jam made from fresh strawberries.
-  "Raspberry Jam", // A tart and sweet jam made with raspberries.
-  "Lemon Curd", // A tangy, sweet curd made with lemons.
-  "Whipped Cream", // A light and fluffy whipped cream filling.
-  "Pastry Cream", // A thick, creamy filling often used in pastries.
-  "Cream Cheese Filling", // A rich, tangy cream cheese filling.
-  "Peanut Butter Filling", // A creamy and nutty peanut butter filling.
-  "Caramel Filling", // A rich, sweet caramel filling.
-  "Mocha Cream", // A filling with a blend of coffee and chocolate flavors.
-  "Orange Marmalade", // A sweet and tangy marmalade filling made from oranges.
-  "Coconut Cream", // A rich, creamy filling made with coconut.
-  "Blueberry Compote", // A fruity and sweet filling made with blueberries.
-  "Nutella", // A creamy hazelnut spread filling for a rich, nutty flavor.
-  "Banana Cream", // A creamy, sweet filling with a banana flavor.
-  "Apple Cinnamon Filling", // A sweet and spiced filling made with apples and cinnamon.
-  "Tiramisu Filling", // A creamy filling inspired by the classic Italian dessert.
-  "Matcha Filling", // A delicate green tea-flavored filling.
-  "Pumpkin Spice Filling", // A spiced filling with pumpkin and cinnamon flavors.
-  "Pineapple Filling", // A sweet and tangy filling made with fresh pineapple.
-]
