+++ src/constants/sweet/cookies.ts
@@ -1,0 +1,91 @@
+// List of cookie flavors categorized into basic and premium flavors
+export const COOKIE_FLAVORS = {
+  basic: [
+    "Chocolate Chip", // A classic favorite with semi-sweet chocolate chips.
+    "Sugar", // A simple, sweet, and crispy sugar cookie.
+    "Peanut Butter", // A rich, nutty cookie made with creamy peanut butter.
+    "Oatmeal Raisin", // A wholesome cookie filled with oats and raisins.
+    "Snickerdoodle", // A soft and chewy cinnamon-sugar coated cookie.
+    "Double Chocolate", // A chocolate lover's dream with cocoa and chocolate chips.
+    "Funfetti", // A cheerful cookie loaded with colorful sprinkles.
+    "Shortbread", // A simple, buttery, and crumbly cookie.
+  ],
+  premium: [
+    "White Chocolate Macadamia Nut", // A smooth cookie with white chocolate and crunchy macadamia nuts.
+    "Gingerbread", // Spicy and aromatic ginger-flavored cookies, often enjoyed during the holidays.
+    "Molasses", // A spiced cookie made with molasses for a deep, sweet flavor.
+    "Lemon", // A zesty and refreshing lemon-flavored cookie.
+    "Pecan Sandies", // A buttery shortbread cookie with chopped pecans.
+    "Pumpkin Spice", // A seasonal favorite with pumpkin and warm spices.
+    "Toffee", // A rich cookie featuring sweet, crunchy toffee bits.
+    "Almond Joy", // A cookie with almonds, coconut, and chocolate.
+    "Coconut Chocolate Chip", // A tropical cookie with coconut and chocolate chips.
+    "Espresso", // A strong, coffee-flavored cookie for coffee lovers.
+    "S'mores", // A fun, gooey cookie with marshmallows, chocolate, and graham cracker crumbs.
+    "Maple Pecan", // A sweet and nutty cookie with maple syrup and pecans.
+    "Brownie", // A fudgy, chocolatey cookie with a brownie-like texture.
+    "Cinnamon Roll", // A warm, cinnamon-spiced cookie inspired by cinnamon rolls.
+  ],
+}
+
+// List of frosting options categorized into basic and premium frostings
+export const COOKIE_FROSTINGS = {
+  basic: [
+    "Buttercream Frosting", // A smooth and sweet frosting made with butter and powdered sugar.
+    "Cream Cheese Frosting", // A tangy, smooth frosting made with cream cheese and sugar.
+    "Chocolate Ganache", // A rich, glossy frosting made from chocolate and cream.
+    "Royal Icing", // A stiff, glossy icing often used for decorating cookies.
+    "Glaze", // A thin, sugary glaze that provides a shiny finish.
+    "Vanilla Frosting", // A classic frosting made with butter, sugar, and vanilla.
+    "Chocolate Frosting", // A rich, chocolatey frosting for any occasion.
+    "Peanut Butter Frosting", // A creamy frosting made with peanut butter for a nutty sweetness.
+  ],
+  premium: [
+    "Caramel Frosting", // A rich, buttery frosting made with caramelized sugar.
+    "Salted Caramel Frosting", // A frosting with a sweet and salty flavor combination.
+    "Mint Frosting", // A fresh and cool frosting with mint flavor.
+    "Lemon Frosting", // A tangy, citrusy frosting with lemon zest.
+    "Coffee Frosting", // A frosting with a subtle coffee flavor.
+    "Marshmallow Frosting", // A fluffy and soft frosting made with marshmallows.
+    "Maple Frosting", // A sweet frosting with the rich flavor of maple syrup.
+    "Coconut Frosting", // A sweet and creamy frosting with coconut flavor.
+    "Strawberry Frosting", // A fruity, sweet frosting made with fresh strawberries.
+    "Raspberry Frosting", // A tangy and sweet frosting made with raspberries.
+    "Orange Frosting", // A citrusy frosting with a bright orange flavor.
+    "Cinnamon Frosting", // A spiced frosting with warm cinnamon flavor.
+    "Toffee Frosting", // A rich frosting with a toffee flavor.
+    "Caramelized Brown Sugar Frosting", // A rich, golden-brown frosting with a caramelized brown sugar taste.
+    "Churro Frosting", // A cinnamon-sugar frosting inspired by churros.
+  ],
+}
+
+// List of possible cookie toppings categorized into basic and premium toppings
+export const COOKIE_TOPPINGS = {
+  basic: [
+    "Sprinkles", // Colorful sprinkles for a fun, festive touch.
+    "Chopped Nuts", // Chopped nuts like almonds, walnuts, or pecans for added crunch.
+    "Mini Marshmallows", // Tiny marshmallows that add a chewy texture and sweetness.
+    "Chocolate Chips", // Classic chocolate chips to enhance any cookie.
+    "Toffee Bits", // Small bits of toffee for a sweet, crunchy addition.
+    "Caramel Drizzle", // A sweet, sticky drizzle of caramel sauce.
+    "Sea Salt", // A sprinkle of sea salt for a sweet and salty flavor contrast.
+    "Coconut Flakes", // Shredded coconut for a tropical flair.
+  ],
+  premium: [
+    "Crushed Oreos", // Crushed Oreo cookies for a chocolatey crunch.
+    "Candy Pieces", // Colorful candy pieces, like M&Ms or Skittles.
+    "Peanut Butter Cups", // Chopped peanut butter cups for a rich, nutty topping.
+    "M&Ms", // Colorful chocolate candies for a fun, crunchy topping.
+    "Bacon Bits", // Crispy bacon bits for a savory-sweet topping.
+    "Chopped Fruit (Cherries, Strawberries, etc.)", // Fresh fruit like cherries or strawberries for a fruity topping.
+    "Shredded Coconut", // Shredded coconut for added texture and sweetness.
+    "Candy Canes", // Crushed candy canes for a festive, minty touch.
+    "Graham Cracker Crumbs", // Crushed graham crackers for a slightly sweet, crunchy topping.
+    "Chocolate Shavings", // Thin shavings of chocolate for an elegant, decorative touch.
+    "Whipped Cream", // Light and fluffy whipped cream for a creamy topping.
+    "Powdered Sugar", // A dusting of powdered sugar for a soft, sweet finish.
+    "Almond Slivers", // Thin slices of almonds for a crunchy, nutty topping.
+    "Cinnamon Sugar", // A sweet and spiced topping made with cinnamon and sugar.
+    "Crushed Pretzels", // Salty pretzels for a sweet-salty topping combination.
+  ],
+}
