+++ src/constants/sweet/cupcakes.ts
@@ -1,0 +1,115 @@
+// Object separating basic and premium cupcake flavors.
+export const CUPCAKE_FLAVORS = {
+  basic: [
+    "Vanilla", // A classic, light, and sweet flavor with a hint of vanilla.
+    "Chocolate", // A rich and decadent chocolate flavor.
+    "Red Velvet", // A soft, moist cake with a hint of cocoa and a vibrant red color.
+    "Carrot", // A spiced cake made with grated carrots and often includes nuts or raisins.
+    "Lemon", // A zesty and refreshing lemon-flavored cupcake.
+    "Strawberry", // A fruity and sweet flavor with fresh strawberry essence.
+    "Funfetti", // A colorful cupcake filled with sprinkles for a fun, festive look.
+    "Coffee", // A rich and robust flavor inspired by coffee beans.
+    "Cookies and Cream", // A sweet flavor with crushed cookies mixed into the batter.
+    "Peanut Butter", // A nutty and creamy cupcake made with peanut butter.
+  ],
+  premium: [
+    "Marble", // A swirl of vanilla and chocolate for a marble effect and dual flavors.
+    "Matcha", // A delicate, earthy flavor made with powdered green tea.
+    "Salted Caramel", // A sweet and salty flavor made with caramel and a touch of salt.
+    "Mocha", // A coffee and chocolate blend for a rich, mocha-inspired flavor.
+    "Coconut", // A tropical flavor with a subtle coconut essence.
+    "Banana", // A moist cupcake with the sweet and fruity taste of bananas.
+    "Almond", // A nutty flavor with a rich, aromatic almond essence.
+    "Chocolate Mint", // A refreshing combination of mint and chocolate.
+    "Raspberry", // A tart and sweet flavor made with fresh raspberries.
+    "Tiramisu", // A coffee-flavored cupcake inspired by the classic Italian dessert.
+    "Mint Chocolate Chip", // A refreshing mint flavor with chocolate chips mixed in.
+    "Maple Bacon", // A savory-sweet flavor with maple syrup and crispy bacon bits.
+    "Churro", // A cinnamon-sugar flavored cupcake inspired by the classic churro.
+  ],
+}
+
+// Object separating basic and premium cupcake frostings.
+export const CUPCAKE_FROSTINGS = {
+  basic: [
+    "Buttercream", // A rich, smooth frosting made with butter, sugar, and vanilla.
+    "Cream Cheese Frosting", // A tangy, smooth frosting made with cream cheese and sugar.
+    "Chocolate Frosting", // A rich, chocolatey frosting for any chocolate cupcake.
+    "Whipped Cream", // A light, fluffy frosting made with whipped cream.
+    "Ganache", // A glossy, rich frosting made from chocolate and cream.
+    "Swiss Meringue Buttercream", // A silky, smooth buttercream made with egg whites, sugar, and butter.
+    "Italian Meringue Buttercream", // A smooth, stable buttercream made with egg whites and sugar syrup.
+    "Peanut Butter Frosting", // A creamy frosting made with peanut butter for a nutty sweetness.
+    "Caramel Frosting", // A rich, sweet frosting made with caramelized sugar.
+    "Salted Caramel Frosting", // A frosting with a sweet and salty flavor combination.
+    "Coconut Frosting", // A smooth frosting with the rich flavor of coconut.
+    "Strawberry Frosting", // A sweet frosting made with fresh strawberries.
+    "Lemon Frosting", // A tangy, citrusy frosting made with lemon zest.
+    "Mint Frosting", // A refreshing frosting with mint flavor.
+    "Coffee Frosting", // A rich, coffee-flavored frosting.
+    "Marshmallow Frosting", // A light, fluffy frosting made with marshmallows.
+    "Mocha Frosting", // A blend of coffee and chocolate in a smooth frosting.
+  ],
+  premium: [
+    "Raspberry Frosting", // A fruity, sweet frosting made with raspberries.
+    "Churro Frosting", // A spiced frosting with cinnamon and sugar, inspired by churros.
+    "Chocolate Hazelnut Frosting", // A creamy, nutty frosting with chocolate and hazelnut flavors.
+  ],
+}
+
+// Object separating basic and premium cupcake toppings.
+export const CUPCAKE_TOPPINGS = {
+  basic: [
+    "Sprinkles", // Colorful sprinkles to add a fun, festive touch.
+    "Chopped Nuts", // Chopped nuts like almonds, walnuts, or pecans for a crunchy texture.
+    "Chocolate Chips", // Classic chocolate chips for extra sweetness.
+    "Caramel Drizzle", // A sweet drizzle of caramel sauce for added richness.
+    "Oreos", // Crushed Oreos for a chocolatey crunch.
+    "Chocolate Shavings", // Thin shavings of chocolate for a decorative finish.
+    "Graham Cracker Crumbs", // Crushed graham crackers for a crunchy topping.
+    "Candy Pieces", // Colorful candy pieces like M&Ms or Skittles for a playful topping.
+    "Peanut Butter Cups", // Chopped peanut butter cups for a rich, nutty topping.
+    "Whipped Cream", // Light and fluffy whipped cream for a creamy topping.
+    "Crushed Cookies", // Crushed cookies like Oreos or chocolate chip cookies for texture.
+    "Butter Toffee", // Sweet, crunchy butter toffee for an indulgent topping.
+  ],
+  premium: [
+    "Mini Marshmallows", // Tiny marshmallows for a soft, chewy topping.
+    "Shredded Coconut", // A tropical topping made with shredded coconut.
+    "Meringue Kisses", // Light, crispy meringue kisses for a delicate topping.
+    "Cotton Candy", // Sweet and fluffy cotton candy for a whimsical touch.
+    "Rainbow Dust", // Shimmering dust for a colorful, sparkling effect.
+    "Candied Ginger", // Sweet, candied ginger for a spicy-sweet topping.
+    "Fresh Strawberries", // Fresh fruit like strawberries for a fruity topping.
+    "Fresh Blueberries", // Fresh fruit like strawberries for a fruity topping.
+    "Fresh Raspberries", // Fresh fruit like strawberries for a fruity topping.
+    "Gold Leaf", // Elegant gold leaf for a luxurious touch.
+    "Pretzels", // Crispy, salty pretzels for a sweet-salty contrast.
+    "Cinnamon Sugar", // A blend of cinnamon and sugar for a sweet, spiced topping.
+  ],
+}
+
+// List of possible cupcake fillings. These fillings can be added to cupcakes for extra flavor and texture.
+export const CUPCAKE_FILLINGS = [
+  "Vanilla Custard", // A creamy, rich custard filling with a vanilla flavor.
+  "Chocolate Ganache", // A thick, rich chocolate ganache filling.
+  "Strawberry Jam", // A sweet and fruity jam made from fresh strawberries.
+  "Raspberry Jam", // A tart and sweet jam made with raspberries.
+  "Lemon Curd", // A tangy, sweet curd made with lemons.
+  "Whipped Cream", // A light and fluffy whipped cream filling.
+  "Pastry Cream", // A thick, creamy filling often used in pastries.
+  "Cream Cheese Filling", // A rich, tangy cream cheese filling.
+  "Peanut Butter Filling", // A creamy and nutty peanut butter filling.
+  "Caramel Filling", // A rich, sweet caramel filling.
+  "Mocha Cream", // A filling with a blend of coffee and chocolate flavors.
+  "Orange Marmalade", // A sweet and tangy marmalade filling made from oranges.
+  "Coconut Cream", // A rich, creamy filling made with coconut.
+  "Blueberry Compote", // A fruity and sweet filling made with blueberries.
+  "Nutella", // A creamy hazelnut spread filling for a rich, nutty flavor.
+  "Banana Cream", // A creamy, sweet filling with a banana flavor.
+  "Apple Cinnamon Filling", // A sweet and spiced filling made with apples and cinnamon.
+  "Tiramisu Filling", // A creamy filling inspired by the classic Italian dessert.
+  "Matcha Filling", // A delicate green tea-flavored filling.
+  "Pumpkin Spice Filling", // A spiced filling with pumpkin and cinnamon flavors.
+  "Pineapple Filling", // A sweet and tangy filling made with fresh pineapple.
+]
